+++ src/client.ts
@@ -2,8 +2,8 @@ import { AWSSignerV4 } from "../deps.ts";
 import type { CreateBucketOptions } from "./types.ts";
 import { S3Error } from "./error.ts";
 import { S3Bucket } from "./bucket.ts";
-import { doRequest, encoder } from "./utils.ts";
-import type { Params } from "./utils.ts";
+import { doRequest, encoder } from "./request.ts";
+import type { Params } from "./request.ts";
 
 export interface S3Config {
   region: string;
